% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSPsurv.R
\name{betas.slice.sampling}
\alias{betas.slice.sampling}
\title{betas.slice.sampling}
\usage{
betas.slice.sampling(
  Sigma.b,
  Y,
  Y0,
  X,
  W,
  betas,
  delta,
  C,
  LY,
  rho,
  w,
  m,
  form
)
}
\arguments{
\item{Sigma.b}{variance estimate of betas}

\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{X}{covariates for betas}

\item{W}{spatial random effects}

\item{betas}{current value of betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling. A vector of values for beta, gamma, rho.}

\item{m}{limit on steps in the slice sampling}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
slice sampling for betas
}
