% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capdist.R
\docType{data}
\name{capdist}
\alias{capdist}
\title{Gleditsch and Ward Distance data}
\format{
A data frame with 41006 rows and 6 variables
}
\source{
...
}
\usage{
data(capdist)
}
\description{
Gleditsch and Ward Distance data
}
\details{
\describe{
   \item{numa}{...}
   \item{ida}{...}
   \item{numb}{...}
   \item{idb}{...}
   \item{kmdist}{...}
   \item{midist}{...}
}
}
\keyword{datasets}
