% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcmcfrailtySP}
\alias{mcmcfrailtySP}
\title{mcmcfrailtySP}
\usage{
mcmcfrailtySP(
  Y,
  Y0,
  C,
  LY,
  X,
  Z,
  S,
  N,
  burn,
  thin,
  w = c(1, 1, 1),
  m = 10,
  form,
  prop.var
)
}
\arguments{
\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{C}{censoring indicator}

\item{LY}{last observation year}

\item{X}{covariates for betas}

\item{Z}{covariates for gammas}

\item{S}{spatial information (e.g. district)}

\item{N}{number of MCMC iterations}

\item{burn}{burn-in to be discarded}

\item{thin}{thinning to prevent from autocorrelation}

\item{w}{size of the slice in the slice sampling for (betas, gammas, rho)}

\item{m}{limit on steps in the slice sampling. A vector of values for beta, gamma, rho.}

\item{form}{type of parametric model (Exponential or Weibull)}

\item{prop.var}{proposal variance for Metropolis-Hastings}

\item{A}{adjacency information corresponding to spatial information}
}
\value{
chain of the variables of interest
}
\description{
Markov Chain Monte Carlo (MCMC) routine to run Bayesian non-spatial frailties split population survival model
}
