% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSPsurv.R
\name{mcmcSP}
\alias{mcmcSP}
\title{mcmcSP}
\usage{
mcmcSP(Y, Y0, C, LY, X, Z, N, burn, thin, w = c(1, 1, 1), m = 10, form)
}
\arguments{
\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{C}{censoring indicator}

\item{X}{covariates for betas}

\item{Z}{covariates for gammas}

\item{N}{number of MCMC iterations}

\item{burn}{burn-in to be discarded}

\item{thin}{thinning to prevent from autocorrelation}

\item{w}{size of the slice in the slice sampling for (betas, gammas, rho)}

\item{m}{limit on steps in the slice sampling. A vector of values for beta, gamma, rho.}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
chain of the variables of interest
}
\description{
Markov Chain Monte Carlo (MCMC) to run Bayesian split population survival model with no frailties
}
