% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rho.post}
\alias{rho.post}
\title{rho.post}
\usage{
rho.post(Y, Y0, eXB, delta, C, LY, rho, a = 1, b = 1)
}
\arguments{
\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{eXB}{exponentiated vector of covariates times betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{LY}{last observation year}

\item{rho}{current value of rho}

\item{a}{shape parameter of gammas prior}

\item{b}{scale parameter of gammas prior}
}
\value{
log- posterior density of betas
}
\description{
log-posterior distribution of rho
}
