% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rho.slice.sampling}
\alias{rho.slice.sampling}
\title{rho.slice.sampling}
\usage{
rho.slice.sampling(
  Y,
  Y0,
  eXB,
  delta,
  C,
  LY,
  rho,
  w,
  m,
  lower = 0.01,
  upper = +Inf
)
}
\arguments{
\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{eXB}{exponentiated vector of covariates times betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{LY}{last observation year}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling. A vector of values for beta, gamma, rho.}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for rho
}
