% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_SA.R
\name{spatial_SA}
\alias{spatial_SA}
\title{spatial_SA}
\usage{
spatial_SA(data, var_ccode, threshold = 800L)
}
\arguments{
\item{data}{data.frame.}

\item{var_ccode}{name of the variable that contains the country codes.}

\item{threshold}{...}
}
\value{
list. Contains database with variable sp_id (S) and matrix A
}
\description{
matrix A and sp_id (S)
}
