% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPstats.R
\name{SPstats}
\alias{SPstats}
\title{SPstats}
\usage{
SPstats(object)
}
\arguments{
\item{object}{An object of the output of pooled, exchangeable, or spatial Split Population survival model .}
}
\value{
List.
}
\description{
Calculates the deviance information criterion (DIC)
     and Log-likelihood for fitted model outputs of pooled,
exchangeable, and spatial Split Population survival models for which a log-likelihood can be obtained using the formula
\emph{DIC = -2 * (L - P)}, where \emph{L} is the log likelihood of the data given the posterior means of the parameter and
\emph{P} is the  estimate of the effective number of parameters in the model.
}
