% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSPsurv.R
\name{univ.gammas.slice.sampling2}
\alias{univ.gammas.slice.sampling2}
\title{univ.gammas.slice.sampling2}
\usage{
univ.gammas.slice.sampling2(
  gammas.p,
  p,
  Sigma.g,
  Y,
  Y0,
  eXB,
  Z,
  V,
  gammas,
  C,
  LY,
  rho,
  w,
  m,
  lower = -Inf,
  upper = +Inf,
  form
)
}
\arguments{
\item{gammas.p}{current value of the pth element of gammas}

\item{p}{pth element}

\item{Sigma.g}{variance estimate of gammas}

\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{V}{spatial random effects}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{rho}{current value of rho}

\item{w}{size of the slice in the slice sampling. A vector of values for beta, gamma, rho.}

\item{m}{limit on steps in the slice sampling}

\item{lower}{lower bound on support of the distribution}

\item{upper}{upper bound on support of the distribution}

\item{form}{type of parametric model (Exponential or Weibull)}
}
\value{
One sample update using slice sampling
}
\description{
univariate slice sampling for gammas.p
}
