% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{V.MH.sampling}
\alias{V.MH.sampling}
\title{V.MH.sampling}
\usage{
V.MH.sampling(S, A, lambda, Y, Y0, eXB, Z, V, gammas, C, LY, rho, prop.var)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{A}{adjacency information corresponding to spatial information}

\item{lambda}{CAR parameter}

\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{eXB}{exponentiated vector of covariates times betas}

\item{Z}{covariates for gammas}

\item{V}{spatial random effects}

\item{gammas}{current value of gammas}

\item{C}{censoring indicator}

\item{LY}{last observation year}

\item{rho}{current value of rho}

\item{prop.var}{proposal variance for Metropolis-Hastings}
}
\value{
One sample update using slice sampling
}
\description{
MH sampling for rcpp_log_dmvnorm
}
