% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{W.F.MH.sampling}
\alias{W.F.MH.sampling}
\title{W.F.MH.sampling (Cure Model with Frailties)}
\usage{
W.F.MH.sampling(Sigma.w, S, Y, Y0, X, W, betas, delta, C, LY, rho, prop.var)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{X}{covariates for betas}

\item{W}{spatial random effects}

\item{betas}{current value of betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{LY}{last observation year}

\item{rho}{current value of rho}

\item{prop.var}{proposal variance for Metropolis-Hastings}
}
\value{
One sample update using slice sampling
}
\description{
MH sampling for W
}
