% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSPsurv.R
\name{W.F.post}
\alias{W.F.post}
\title{W.F.post}
\usage{
W.F.post(Sigma.w, S, Y, Y0, X, W, betas, delta, C, LY, rho)
}
\arguments{
\item{S}{spatial information (e.g. district)}

\item{Y}{the time (duration) dependent variable for the survival stage (t)}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{X}{covariates for betas}

\item{W}{spatial random effects}

\item{betas}{current value of betas}

\item{delta}{probability of true censoring}

\item{C}{censoring indicator}

\item{rho}{current value of rho}
}
\value{
log- posterior density of W
}
\description{
log-posterior distribution of W with sth element fixed as W.s
}
