% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capdist.R
\docType{data}
\name{capdist}
\alias{capdist}
\title{Gleditsch and Ward Distance data}
\format{
A data frame with 41006 rows and 6 variables
}
\source{
Gleditsch, Kristian S., and Michael D. Ward. (2001). "Measuring Space: A Minimum-Distance Database and Applications to International Studies." Journal of Conflict Resolution 38(6): 739-758.
}
\usage{
data(capdist)
}
\description{
Dyadic dataset extracted from \href{t.ly/bGE9}{Gleditsch and Ward (2001)}.
The dataset contains information on the distace between capital cities among independent nation-states.
}
\details{
\describe{
   \item{numa}{COW code -- country A}
   \item{ida}{Three letter ISO code -- country A}
   \item{numb}{COW code -- country B}
   \item{idb}{Three letter ISO code -- country B}
   \item{kmdist}{...}
   \item{midist}{...}
}
}
\keyword{datasets}
