% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialSPsurv.R
\name{spatialSPsurv}
\alias{spatialSPsurv}
\title{spatialSPsurv}
\usage{
spatialSPsurv(
  duration,
  immune,
  Y0,
  LY,
  S,
  A,
  data,
  N,
  burn,
  thin,
  w = c(1, 1, 1),
  m = 10,
  form = c("Weibull", "exponential", "loglog"),
  prop.var
)
}
\arguments{
\item{duration}{...}

\item{immune}{...}

\item{Y0}{the elapsed time since inception until the beginning of time period (t-1)}

\item{LY}{last observation year}

\item{S}{spatial information (e.g. district ID) for each observation that matches the spatial matrix row/column information}

\item{A}{Spatial Matrix (load separate spatial weights matrix file)}

\item{data}{...}

\item{N}{number of MCMC iterations}

\item{burn}{burn-in to be discarded}

\item{thin}{thinning to prevent from autocorrelation}

\item{w}{size of the slice in the slice sampling for (betas, gammas, rho). Write it as a vector. E.g. c(1,1,1)}

\item{m}{limit on steps in the slice sampling. A vector of values for beta, gamma, rho.}

\item{form}{type of parametric model (Exponential or Weibull)}

\item{prop.var}{proposal variance for Metropolis-Hastings}
}
\value{
chain of the variables of interest
}
\description{
Markov Chain Monte Carlo (MCMC) to run Bayesian spatial split population survival model
}
